requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.20.0
ultralytics>=8.0.0
Pillow>=8.3.0
plotly>=5.0.0
     
